Included validateLink into utils
